moved to bearer token auth added api key option
